Change JSON output format for easy parsing
Fixes #4

The previous JSON output format depends on the named attributes and
block_types. It's hard to parse. If we use a slice instead of a map, we
can make the key name predictable. However, I don't want to change the
internal data structure for the convenience of the only JSON format. I
want to also use a tfschema as a Go library not only CLI. So, the data
model of tfschema should be separated from the requirements of the
output format.

Extract format related code from the `tfschema` package to a new
`formatter` package, define a new type for output and call it from
`command` package, and change the JSON output format.
In order to reduce the possibility of future conflicts, the name of keys
are borrowed from the schema definition of the grpc provider's proto
file.