Change type notation to HCL2 type annotation
Fixes #9

Terraform v0.12 introduced a new `SchemaConfigModeAttr` feature.
Most attributes have simple types, but if `SchemaConfigModeAttr` is set for
an attribute, it is syntactically NestedBlock but semantically interpreted
as an Attribute. In this case, Attribute has a complex data type. It is
reasonable to use the same notation as the type annotation in HCL2 to
represent the correct data type. However, it seems that HCL2 has a type
annotation parser but no writer, so we implement it by ourselves.

See also:
- https://github.com/minamijoyo/tfschema/issues/9
- https://github.com/terraform-providers/terraform-provider-aws/pull/8187
- https://github.com/hashicorp/terraform/pull/20626
- https://www.terraform.io/docs/configuration/types.html

This is a breaking change caused by the upstream,
so I think it is impossible to maintain tfschema backward compatibility.
I tried to be as consistent as possible.