Support Terraform v0.13
Fixes #20

Terraform v0.13 changed the default location for storing provider's
binary to support downloading community providers.
We should support it.

If we could import internal packages in hashicorp/terraform,
it would be implemented easily.
As a compromise, we peek a lock file (.terraform/plugins/selections.json)
and build a path of plugin directories.
It doesn't implement exactly the same logic as Terraform v0.13,
but should be fine for most use cases.